make timer.t pass on solaris under arc unit
Digging into this showed that our `Pv%s%p recursive format expansion
stuff wasn't quite right on solaris.

Introduce a macro to correctly(?) pass the va_list down.  I think
this is a bit sketchy but seems to work out ok.  I wouldn't be
surprised if this needs revisiting again in the future.