Add thread pools
This is a pretty big commit due to a bunch of tuning and
experimentation guided by some long cache profiling runs.

More details on the thread pool implementation can be found in both
job.c and job.h.

tpool.t is able to sustain a rate in excess of 40mm jobs per second per
core with a linear relationship to the number of physical cores.
We observe this trailing off as virtual cores are used; results were
observed on a 16 core (32 with HT enable) system.  Performance rapidly
reduced to half the jobs-per-second-per-core once we passed 16 cores.

* Reduced contention for ph_counter.  We no longer use ck_sequence
  for counters as it introduced a number of expensive atomics that
  we don't strictly need in our use case.  We still follow the
  spirit of the sequence operation but the strength of the published
  counters is a bit "weaker" in return for increased throughput.

* Replaced ph_time_t with a regular timeval.  This resolves some
  resolution issues with time math (signed 64-bit was not enough).

* Expended ph_log() to include a higher resolution timestamp,
  a readable level label and intelligently handle trailing newlines.
  In addition, we print the thread name and tid.

* Runs clean under valgrind --leak-check=full.  This is managed using
  atexit() to force destroy a number of key structures that wouldn't
  otherwise get destroyed.

* ph_num_cores() - returns the number of physical cores (not
  hyperthreads).  We use this to size the scheduler pool size and also
  a pool size in the tpool.t test case.