simplify timer and nbio jobs
I didn't like the owner and versioning thing; it was too complex.
This simplifies timer processing by making the following observations:

Most timers on IO jobs are for timeout events, and those are typically
far enough in the future that repeated remove/insert operations on
an active fd are likely to compute the same bucket in the wheel.

In addition, timeout triggers are pretty rare compared to actual IO
dispatches anyway.

Rather than force the sequence:

```
descriptor_ready:
  lock_wheel()
  remove_timer()
  unlock_wheel()
  dispatch_event()
  lock_wheel()
  add_timer()
  unlock_wheel()
```

we can use:

```
descriptor_ready:
  disable_timer()
  dispatch_event()
  enable_timer()
```

where the disable_timer function is a single CAS operation and needn't
block any other dispatching threads.  This minimizes wheel locks to
the timer dispatcher itself, or situations where the timer is newly
added, being completely removed, or is re-bucketed.

As another nod to avoid racy situations, this adds an additional
before_dispatch callback in the timer_tick routine that allows the
driver of the wheel to opt out of dispatching just-in-time or, as
wired up in nbio.c, disable the IO portion of the event in the kernel.