Remove doxygen thingy, replace with markdown extraction
I don't really like the very bitty and disjoint nature of doxygen,
so here's how we're doing docs:

By convention, each header file begins with a section of markdown
text enclosed in a C-style docblock region (enclosed in /** */).
This is intended to describe in broad terms the facilities provided
by that file and give some general examples.

Then, each declaration in the header is prefixed with a docblock
comment that contains documentation for that declaration.

The convention for these is that the first line of the docblock is
the abbreviated summary description of the thing and the remainder
document it in more detail.

The mkdoc.php script walks over the header files and produces a series
of reasonably formatted markdown files in the docs dir which you are
then free to consume using tools of your choice.

The end result should be something reasonably well consumable by
humans.