Combine epoch and counter heads into one struct
This saves a TLS key and "solves" the TODO for recycling heads by
having that handled by the epoch recycling functionality.

Remove the barrier calls in the TLS dtor as these can cause deadlock
on scheduler shutdown, and the memory associated with them should be
sufficiently small that we don't really care about getting that
back any time soon; much rather leave it hanging around for another
thread to recycle and use later, assuming that this thread spawn/exit
pattern is common in the applications that opt to use it, even if
we'd prefer that the application use a set of long-lived threads anyway.