Add ph_string_XXX suite of functions
Docs are inline in the header file.

Took the opportunity to increase the memtype id of the first valid
memtype to make it easier to implement some of the string functions,
as we are slightly evil and abuse negative memtype values.  Since 0
doesn't have a valid negative value, we need to ensure that the first
memtype is not zero.  We also want to avoid flipping the invalid memtype
value (-1) to 1, so our first valid memtype value is 2.

Realized that we might as well just redefine PH_MEMTYPE_INVALID to 0.