address https://github.com/facebook/libphenom/issues/6
What was going on here was a bit tricky; I'm not entirely sure why
this only triggered with clang on OSX.

I could reproduce a very similar stack trace reasonably often with
dns.t

The sequence of operations was this:

A DNS worker thread finished:

- the counter TLS key was destroyed, tearing down the epoch_record
  associated with that thread
- the ph_thread TLS key was destroyed.  There were deferred items
  to be reclaimed (scheduled with ph_thread_epoch_defer) and these
  were dispatched now in the TLS dtor.  This in turn called
  ph_mem_free(), which in turn wants to update memory counters.
- The counter subsystem noticed that it didn't have a counter head
  for this thread (we tore it down already) so attempted to recycle
  and register a new head.  At some point in this process, the
  memory in the ph_thread_t for the current got corrupted.

We can avoid this by having our spawned threads trigger an explicit
barrier before they exit; this will cause the deferred memory to
get freed and counters updated before we land in the TLS dtor code.

We might just be dodging a problem here; if the dtors allocate memory
we'll also trigger code to establish counter heads.  We'll probably
revisit this issue again in the future.