more work towards https://github.com/facebook/libphenom/issues/6
I could only get this to repro with clang built using homebrew,
and it triggered rarely.

This is a continuation of the prior fix attempt, which "fixes"
things by not triggering the barrier call from inside the TLS
dtor.

This introduces the possibility of a resource leak on thread shutdown
if the thread is not one spawned by the phenom thread functions
directly, so we'll need to do some follow-up work on this.