Add OpenSSL support to ph_sock_t
Addresses https://github.com/facebook/libphenom/issues/20

Since SSL carries around a lot of state, we chose to implement the
bulk of the logic in ph_sock_t.

This commit adds some connectors for BIO->ph_stream_t and so on;
these exist to facilitate the ph_sock_t implementation and have
some limitations (such as missing read primitives) in some cases.

We choose to provide just enough infrastructure for you to build
your protocol on top of ph_sock_t.  In particular, you need to provide
the SSL context and object management.

We may evolve this over time to add more convenience functions.

This commits augments the echo example to demonstrate how to enable
SSL for a listener, and adds a very simple sclient example that shows
how to build an SSL client.

Also: fixup an issue with bufq's where gc can leave them completely
empty of spare buffers.