fixup doc entry name matching
Adding `PH_` for macros meant that struct ph_job was being mislabeled
as a PH_STAILQ_ENTRY.  Make the title check prefer `struct`s that start
the line.  This duplicates a later rule that matches `struct` anywhere
in the line because it is late and I'm too lazy to check to see what
might be affected if I remove the latter.