add helper for getting at a c-string rendition of a ph_string_t
Docs explain it; we prefer to copy (or not copy if possible) to the
stack over using the heap.  The rationale is that most of the strings
we'll pass to c-string APIs will be small enough that stack will be
appropriate; my immediate use case is passing strings to posix APIs
for use as filenames.