augment thread name in ph_log
In environments where we're mixing libphenom with other systems (such as
thrift dispatchers), we may invoke logging from a thread that doesn't
have phenom TLS.  In that case we want to print out a more meaningful
name.

This is two-pronged: when setting up phenom TLS, try to copy the
name from the system.

When logging, if we don't have phenom TLS, use the name and the thread
id.  We try to get the LWPid from the system if we can.