make sure we remove our timer when we free a job
We've seen a couple of timer related faults; on further inspection,
we weren't ever explicitly removing a timer when we freed a job.

This amends ph_job_free to disable a timer at the point where the
free is deferred, and then remove the item from the wheel during the
dtor (this latter is a blocking operation).