Add ph_nbio_queue_affine_func(), ph_job_wakeup()
This reinstate some of the affine dispatch functionality from an earlier
incarnation of phenom.  This allows executing an application defined
callback function in the context of a specific emitter thread (selected
using the emitter_affinity value).

A trivial example of its use is seen in examples/sclient.c where we
read a line from stdin, mutate the state associated with a sock object,
and then wake it up so that it can send out that data.

Another use case is scheduling a periodic job, but causing it to
trigger immediately in response to some other stimulus.