fix size calculation for hash tables
Addresses https://github.com/facebook/libphenom/issues/44

The problem was that the compiler was producing int32_t sized results
when figuring out sizes and offsets.  Combine this with a large number
of elements (large enough to overflow 32-bits) and the table would be
allocated smaller than we needed.

This solves the problem in the fewest keystrokes by promoting the
size of a couple of fields to 64-bit, but has the side effect of
changing the size of the hashtable struct.