For short writev results, update bufq rpos correctly
We were moving rpos to the end unconditionally.  Is situations where
the data is large and a short writev result is encountered, we'd believe
that we had sent all the data that was there when we actually had not.

Manifestations would be missing chunks of data in the worst case, or
stalled sessions in the best case (due to protocol mismatch)