Downgrade ssl errors to INFO level
A surprising number of SSL/TLS enabled services on the public internet
have broken configurations (ranging from invalid certs, to "bad" cipher
selections).

Avoid making noise in the log stream when these are encountered.

We probably should have a way to tap into the underlying SSL error stack
in case there are interesting things that could be done at a high layer
in the stack, but for most practical purposes all we care about is that
an error was found.