Allow listeners to opt out of SSL_CTX_free for ph_sock_t
Summary: SSL server contexts are relatively more expensive to create
than client contexts because they have to load certificates, keys and
perhaps also DH parameters for PFS.  The unilateral SSL_CTX_free in
sock_dtor forces a more expensive (and likely blocking!) operation for
these sorts of contexts.

This diff allows disabling the the SSL_CTX_free; when disabled, the
application is solely responsible for managing the lifetime of the
SSL_CTX associated with the SSL object that was enabled on the sock
object.

I've modified `examples/echo.c` to show how to use it in this fashion
for listeners.  The `sclient` example code remains unchanged.

Addresses https://github.com/facebook/libphenom/issues/58