linting: Default to pwd if the .git dir is missing
The cause of the build failure in https://github.com/facebook/libphenom/issues/72
is that the brew build checks out the repo without the .git dir.

The linter is looking to make paths relative to the root, and it was
finding the .git dir to determine where the root was.  Without the git
dir being present, it ends up being relative to `/` and fails the build
because the include guards don't match that scheme.

If the .git dir is missing, we'll now just assume that we're being run
from the root, which is the only way that this script is run in
libphenom.